refactor: moved getter function to helpers
