+++ src/decorators/Getter.ts
@@ -1,5 +1,6 @@
 import { Getter } from 'vuex';
-import { stores, initStore, getClassName, config } from './utils';
+import { stores, initStore, getClassName } from './utils';
+import { GetterFn } from './helpers';
 
 export default function Getter<T, R>(
   target: T,
@@ -17,36 +18,4 @@ export default function Getter<T, R>(
     targetModule,
     (target as any)[key]
   );
-}
-
-export function GetterFn<S, R>(targetModule: any, getterFn: () => {}) {
-  return (state: S, getters: any, rootState: R, rootGetters: any) => {
-    let output;
-    if (config.store && targetModule._caller) {
-      targetModule[`${targetModule._caller}staticGetters`].$store = {
-        state,
-        getters,
-        rootState,
-        rootGetters,
-      };
-      output = getterFn.call(
-        targetModule[`${targetModule._caller}staticGetters`]
-      );
-    } else if (config.store && targetModule._staticGetters) {
-      targetModule._staticGetters.$store = {
-        state,
-        getters,
-        rootState,
-        rootGetters,
-      };
-      output = getterFn.call(targetModule._staticGetters);
-    } else {
-      const thisObject = { $store: { state, getters, rootState, rootGetters } };
-      for (const key of Object.keys(state)) {
-        Object.assign(thisObject, { [key]: (state as any)[key] });
-      }
-      output = getterFn.call(thisObject);
-    }
-    return output;
-  };
 }
